Always separate on spaces and unprintable characters
Some code is sending our backend mentions with a zero-width joiner for
whatever reason, for example:

    @example\u200d skldjg

The mention that would would get picked up is `@example\u200d`, instead
of the expected `@example`.

This patch always treats spaces and unprintable characters as
a terminator. I'm not sure if it ever makes sense to use a space inside
a mention?

Alternatively, I could add a callback function, or something?
Maintaining a list of all space/unprintable characters doesn't strike me
as a good idea, and the `unicode.Is*()` functions contain some
performance optimisations for the common (non-multibyte) use cases.

The downside is some performance loss:

master:

    BenchmarkGetTags-4       2000000               988 ns/op

this patch:

    BenchmarkGetTags-4       1000000              1259 ns/op